Move ContainerExtensions to top level namespace
